Loads of changes.
New Class to run the game and do the drawing, moving etc.
Currently has move, with bounds to not go off the screen.
No Scaling.
Game grid is no longer fit to the screen, and is centered on start.
Can drag the image without registering a touch event to draw a line.